fix modals
